feat: refactor algorithm interface to abstract class
Converted the algorithm interface to an abstract class. Added methods for calculating distances and paths, including `getCompleteDistance` and `getDistance`. Introduced a private static constant for the salesman index and updated distance handling logic.
