+++ src/main/java/fr/iut/pathpilotapi/algorithme/BruteForce.java
@@ -16,24 +16,9 @@ import java.util.stream.IntStream;
  * <p>
  * It's a naive algorithm that will try every possible path to find the best one.
  */
-public class BruteForce implements Algorithme {
-
-    public static final int SALESMAN_INDEX = 0;
-    private List<List<Double>> distances;
-    private List<Integer> bestPath;
-    private double bestDistance;
+public class BruteForce extends Algorithme {
 
     public BruteForce() {
-    }
-
-    /**
-     * Set the matrix of distances between the clients and the salesman.
-     *
-     * @param distances square matrix with the distances between the clients and the salesman
-     */
-    @Override
-    public void setMatrixLocationsRequest(List<List<Double>> distances) {
-        this.distances = distances;
     }
 
     @Override
@@ -43,19 +28,6 @@ public class BruteForce implements Algorithme {
         // List of every number between 1 and the number of clients (distances.size() - 1)
         List<Integer> remainingClientsIndex = IntStream.range(1, distances.size()).boxed().toList();
         findBestPathForItinerary(remainingClientsIndex);
-    }
-
-    @Override
-    public List<Integer> getBestPath() {
-        return bestPath;
-    }
-
-    @Override
-    public Double getDistanceBestPath() {
-        if (getBestPath().isEmpty()) {
-            return 0.0;
-        }
-        return getCompleteDistance(bestPath);
     }
 
     /**
@@ -73,50 +45,6 @@ public class BruteForce implements Algorithme {
                 bestPath = path;
             }
         }
-    }
-
-    /**
-     * Get the distance of a path.
-     * <p>
-     *     The path is a list of clients' index.
-     *     <br>
-     *     The salesman shouldn't be in the path, but we add it to the beginning and the end of the path.
-     * @param bestClientPath the path to calculate the distance
-     * @return the distance of the path
-     */
-    private Double getCompleteDistance(List<Integer> bestClientPath) {
-        Double distance = getDistance(SALESMAN_INDEX, bestClientPath.getFirst());
-        distance += getDistance(bestClientPath);
-        distance += getDistance(bestClientPath.getLast(), SALESMAN_INDEX);
-        return distance;
-    }
-
-    /**
-     * Get the distance of a path.
-     * <p>
-     *     The path is a list of clients' index.
-     *     <br>
-     *     The salesman shouldn't be in the path.
-     * @param bestClientPath the path to calculate the distance
-     * @return the distance of the path
-     */
-    private Double getDistance(List<Integer> bestClientPath) {
-        Double distance = 0.0;
-        for (int i = 1; i < bestClientPath.size() - 1; i++) {
-            distance += getDistance(bestClientPath.get(i), bestClientPath.get(i + 1));
-        }
-        return distance;
-    }
-
-    /**
-     * Get the distance between two clients.
-     *
-     * @param from the index of the first client we want to start from
-     * @param to   the index of the second client we want to go to
-     * @return the distance between the two clients
-     */
-    private Double getDistance(int from, int to) {
-        return distances.get(from).get(to);
     }
 
     public Set<List<Integer>> getCombinaisons(List<Integer> list) {
